Cleaned up the code and separated each components.
