+++ index.html
@@ -10,6 +10,6 @@
 
   <body>
     <div id="root">Not Rendered</div>
-    <script type="module" src="/App.js"></script>
+    <script type="module" src="src/App.js"></script>
   </body>
 </html>
