+++ src/components/Footer.js
@@ -1,0 +1,18 @@
+// Footer component for footer section
+const Footer = () => {
+  return (
+    <div className="footer">
+      Created By
+      <i class="fa-solid fa-heart"></i>
+      <a href="https://www.linkedin.com/in/shikharbansal007/" target="_blank">
+        Shikhar Bansal
+      </a>
+      <i class="fa-solid fa-copyright"></i>2023
+      <strong>
+        Food<span>Fire</span>
+      </strong>
+    </div>
+  );
+};
+
+export default Footer;
