+++ src/config.js
@@ -1,53 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-// import FoodFireLogo from "../Images/Food Fire Logo.png";
-
-/* My Food App structure will look like this, 
-            1) Header
-                - Logo
-                - Nav Items(right side)
-                - Cart
-            2) Body
-                - Search bar
-                - Restaurants List
-                    - Restaurant card
-                        - Image
-                        - Name
-                        - Rating
-            3) Footer
-                - Links
-                - Copyrights
-       
-*/
-
-// Title component for display logo
-const Title = () => (
-  <a href="/">
-    {/* <img className="logo" src={FoodFireLogo} alt="Food Fire Logo" /> */}
-  </a>
-);
-
-// Header component for header section: Logo, Nav Items
-const Header = () => {
-  return (
-    <div className="header">
-      <Title />
-      <div className="nav-items">
-        <ul>
-          <li>Home</li>
-          <li>About</li>
-          <li>Contact</li>
-          <li>
-            <i class="fa-solid fa-cart-shopping"></i>
-          </li>
-        </ul>
-      </div>
-    </div>
-  );
-};
+export const IMG_CDN_URL =
+  "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/";
 
 // RestaurantList is JSON Data for displaying cards
-const restaurantList = [
+export const restaurantList = [
   {
     type: "restaurant",
     data: {
@@ -1008,79 +963,3 @@ const restaurantList = [
     subtype: "basic",
   },
 ];
-
-// Restaurant card component: Image, name, cuisine
-const RestaurantCard = ({
-  cloudinaryImageId,
-  name,
-  cuisines,
-  area,
-  lastMileTravelString,
-  costForTwoString,
-  avgRating,
-}) => {
-  return (
-    <div className="card">
-      <img
-        src={
-          "https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/" +
-          cloudinaryImageId
-        }
-      />
-      <h2>{name}</h2>
-      <h4>{cuisines.join(", ")}</h4>
-      <h4>{area}</h4>
-      <span>
-        <h4>
-          <i class="fa-solid fa-star"></i>
-          {avgRating}
-        </h4>
-        <h4>{lastMileTravelString}</h4>
-        <h4>{costForTwoString}</h4>
-      </span>
-    </div>
-  );
-};
-
-// Body Component for body section: It contain all restaurant cards
-// We are mapping restaurantList array and passing data to RestaurantCard component as props with unique key as index
-const Body = () => {
-  return (
-    <div className="restaurant-list">
-      {restaurantList.map((restaurant) => {
-        return <RestaurantCard key={restaurant.data.id} {...restaurant.data} />;
-      })}
-    </div>
-  );
-};
-
-// Footer component for footer section
-const Footer = () => {
-  return (
-    <div className="footer">
-      Created By
-      <i class="fa-solid fa-heart"></i>
-      <a href="https://www.linkedin.com/in/chetannada/" target="_blank">
-        Chetan Nada
-      </a>
-      <i class="fa-solid fa-copyright"></i>2023
-      <strong>
-        Food<span>Fire</span>
-      </strong>
-    </div>
-  );
-};
-
-// AppLayout component to show: Header, Body, Footer
-const AppLayout = () => {
-  return (
-    <React.Fragment>
-      <Header />
-      <Body />
-      <Footer />
-    </React.Fragment>
-  );
-};
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<AppLayout />);
